address PR comments part 1
