+++ arbor/CMakeLists.txt
@@ -124,12 +124,12 @@ endif()
 add_library(arbor ${arbor_sources} ${arbor_mechanism_sources})
 add_dependencies(arbor build_all_mods)
 
+target_link_libraries(arbor PRIVATE arbor-private-deps arbor-private-headers)
+target_link_libraries(arbor PUBLIC arbor-public-deps arbor-public-headers)
+
 if(ARB_WITH_CUDA_CLANG)
-    target_link_libraries(arbor PRIVATE arbor-private-deps arbor-private-headers ${CUDA_LIBRARIES})
-    target_link_libraries(arbor PUBLIC arbor-public-deps arbor-public-headers ${CUDA_LIBRARIES})
-else()
-    target_link_libraries(arbor PRIVATE arbor-private-deps arbor-private-headers)
-    target_link_libraries(arbor PUBLIC arbor-public-deps arbor-public-headers)
+    target_link_libraries(arbor PRIVATE ${CUDA_LIBRARIES})
+    target_link_libraries(arbor PUBLIC  ${CUDA_LIBRARIES})
 endif()
 
 set_target_properties(arbor PROPERTIES CUDA_RESOLVE_DEVICE_SYMBOLS ON)
