+++ arbor/backends/gpu/matrix_fine.hpp
@@ -47,7 +47,7 @@ void solve_matrix_fine(
     const level_metadata* level_meta,      // information pertaining to each level
     const fvm_index_type* level_lengths,   // lengths of branches of every level concatenated
     const fvm_index_type* level_parents,   // parents of branches of every level concatenated
-    const fvm_index_type* block_index,     // start index (exclusive) into levels for each hip block
+    const fvm_index_type* block_index,     // start index (exclusive) into levels for each gpu block
     fvm_index_type* num_cells,             // the number of cells packed into this single matrix
     fvm_index_type* padded_size,           // length of rhs, d, u, including padding
     unsigned num_blocks,                   // number of blocks
