+++ modcc/printer/cexpr_emit.cpp
@@ -86,7 +86,7 @@ void CExprEmitter::visit(AssignmentExpression* e) {
 void CExprEmitter::visit(PowBinaryExpression* e) {
     out_ << "pow" << '(';
     e->lhs()->accept(this);
-    out_ << ", (double)";
+    out_ << ", (fvm_value_type)";
     e->rhs()->accept(this);
     out_ << ')';
 }
