+++ modcc/printer/gpuprinter.cpp
@@ -64,7 +64,7 @@ std::string emit_gpu_cpp_source(const Module& module_, const printer_options& op
     io::pfxstringstream out;
 
     net_receive && out <<
-        "#ifdef ARB_HAVE_HIP\n"
+        "#ifdef ARB_HIP\n"
         "#include <hip/hip_runtime.h>\n"
         "#endif\n"
         "#include <" << arb_private_header_prefix() << "backends/event.hpp>\n"
